move `process_one_mol` test
