+++ tests/test_gc_meox.py
@@ -1,6 +1,6 @@
 import pytest
 
-from gc_meox_tms import add_derivatization_groups, is_derivatized, remove_derivatization_groups
+from gc_meox_tms import add_derivatization_groups, is_derivatized, remove_derivatization_groups, process_one_mol
 from rdkit import Chem
 
 
@@ -106,3 +106,19 @@ def test_add_derivatization_groups_from_mol(derivatization_groups_data):
     derivatized_smiles = Chem.MolToSmiles(derivatized, kekuleSmiles=True)
 
     assert derivatized_smiles in [expected, alternative]
+
+
+@pytest.mark.parametrize("smiles, expected", [
+    ("CC(N)=O", {"CC(N)=O", "CC(=O)N([Si](C)(C)C)[Si](C)(C)C", "CC(=O)N[Si](C)(C)C"}),
+    ("C[Si](C)(C)OC1=CC=C(O)C=C1", {"OC1=CC=C(O)C=C1", "C[Si](C)(C)OC1=CC=C(O[Si](C)(C)C)C=C1", "C[Si](C)(C)OC1=CC=C(O)C=C1"}),
+    ("CCC(C)=O", {"CCC(C)=O", "CCC(C)=NOC"}),
+    ("CC=NOC", {"CC=O", "CC=NOC"})
+])
+def test_process_one_mol(smiles, expected):
+    """Test processing one molecule."""
+    mol = (smiles, Chem.MolFromSmiles(smiles))
+    n = 40
+    actual = process_one_mol((mol, n))
+    actual = {actual[0], actual[1], *actual[2]}
+
+    assert actual == expected
