+++ tests/test_utils.py
@@ -89,19 +89,3 @@ def test_writing_tsv_content(data, tmp_path):
 
     assert len(lines) == 5
     assert lines[0] == "orig\tderiv. removed\tderiv. added ...\n"
-
-
-@pytest.mark.parametrize("smiles, expected", [
-    ("CC(N)=O", {"CC(N)=O", "CC(=O)N([Si](C)(C)C)[Si](C)(C)C", "CC(=O)N[Si](C)(C)C"}),
-    ("C[Si](C)(C)OC1=CC=C(O)C=C1", {"OC1=CC=C(O)C=C1", "C[Si](C)(C)OC1=CC=C(O[Si](C)(C)C)C=C1", "C[Si](C)(C)OC1=CC=C(O)C=C1"}),
-    ("CCC(C)=O", {"CCC(C)=O", "CCC(C)=NOC"}),
-    ("CC=NOC", {"CC=O", "CC=NOC"})
-])
-def test_process_one_mol(smiles, expected):
-    """Test processing one molecule."""
-    mol = (smiles, MolFromSmiles(smiles))
-    n = 40
-    actual = process_one_mol((mol, n))
-    actual = {actual[0], actual[1], *actual[2]}
-
-    assert actual == expected
